requirements.txt
psycopg2-binary==2.9.9
faker==24.0.0
great-expectations==0.18.12
pytest==8.1.0
pytest-mock==3.14.0
pytest-cov==5.0.0
python-dotenv==1.0.1
sqlalchemy==2.0.29
pandas==2.2.1